rename, add Until(String)
