+++ failgood/jvm/src/failgood/Ignored.kt
@@ -1,6 +1,8 @@
 package failgood
 
 import java.time.Instant
+import java.time.LocalDate
+import java.time.ZoneOffset
 
 /**
  * decide if a test should be ignored. you can add your own matchers like
@@ -8,18 +10,20 @@ import java.time.Instant
  * val onCi = IsIgnored { System.getenv("CI") != null }
  * ```
  */
-fun interface IsIgnored {
+fun interface Ignored {
     fun isIgnored(): Boolean
 
-    object Always : IsIgnored {
+    object Always : Ignored {
         override fun isIgnored(): Boolean = true
     }
 
-    object Never : IsIgnored {
+    object Never : Ignored {
         override fun isIgnored(): Boolean = false
     }
 
-    class Until(private val instant: Instant) : IsIgnored {
+    class Until(private val instant: Instant) : Ignored {
+        constructor(dateString: String) : this(LocalDate.parse(dateString).atStartOfDay().toInstant(ZoneOffset.UTC))
+
         override fun isIgnored(): Boolean {
             return instant.isAfter(Instant.now())
         }
