+++ failgood/jvm/test/failgood/pitest/FailGoodTestUnitFinderTest.kt
@@ -1,6 +1,6 @@
 package failgood.pitest
 
-import failgood.IsIgnored.Always
+import failgood.Ignored.Always
 import failgood.Test
 import failgood.describe
 import failgood.mock.mock
